moved CloneOptions and IGitAccount out to separate modules
