+++ app/src/lib/git/authentication.ts
@@ -1,17 +1,7 @@
 import * as Path from 'path'
 
 import { GitError as DugiteError } from 'dugite'
-
-/**
- * An account which can be used to potentially authenticate with a git server.
- */
-export interface IGitAccount {
-  /** The login/username to authenticate with. */
-  readonly login: string
-
-  /** The endpoint with which the user is authenticating. */
-  readonly endpoint: string
-}
+import { IGitAccount } from '../../models/git-account'
 
 /** Get the environment for authenticating remote operations. */
 export function envForAuthentication(auth: IGitAccount | null): Object {
