+++ app/src/lib/git/revert.ts
@@ -1,9 +1,10 @@
 import { git, gitNetworkArguments, IGitExecutionOptions } from './core'
-import { envForAuthentication, IGitAccount } from './authentication'
+import { envForAuthentication } from './authentication'
 
 import { Repository } from '../../models/repository'
 import { Commit } from '../../models/commit'
 import { IRevertProgress } from '../../models/progress'
+import { IGitAccount } from '../../models/git-account'
 
 import { executionOptionsWithProgress } from '../progress/from-process'
 import { RevertProgressParser } from '../progress/revert'
